Application/requirements.txt
langchain 
openai 
transformers 
typing 
python-dotenv 
google-search-results 
splunk-sdk 
tiktoken 
PyPDF2 
faiss-cpu
streamlit
bs4